Move polygon validation code into service
