+++ src/app/app.component.ts
@@ -1,12 +1,10 @@
 import { Component, OnInit } from '@angular/core';
 import { HttpParams } from '@angular/common/http';
-import { MatSnackBar, MatBottomSheet } from '@angular/material';
+import { MatBottomSheet } from '@angular/material';
 
 import { Store } from '@ngrx/store';
 
-import {
-  filter, map, switchMap, tap, catchError
-} from 'rxjs/operators';
+import { filter, map, switchMap, tap, catchError } from 'rxjs/operators';
 
 import { SpreadsheetComponent } from '@components/spreadsheet';
 
@@ -32,16 +30,14 @@ export class AppComponent implements OnInit {
 
   constructor(
     private store$: Store<AppState>,
-    private snackBar: MatSnackBar,
     private bottomSheet: MatBottomSheet,
     private mapService: services.MapService,
-    private asfApiService: services.AsfApiService,
     private urlStateService: services.UrlStateService,
-    private wktService: services.WktService,
+    private polygonValidationService: services.PolygonValidationService,
   ) {}
 
   public ngOnInit(): void {
-    this.validateSearchPolygons();
+    this.polygonValidationService.validate();
   }
 
   public onLoadUrlState(): void {
@@ -62,51 +58,5 @@ export class AppComponent implements OnInit {
     this.store$.dispatch(new granulesStore.ClearGranules());
     this.store$.dispatch(new filterStore.ClearFilters());
     this.mapService.clearDrawLayer();
-  }
-
-  private validateSearchPolygons(): void {
-    this.mapService.searchPolygon$.pipe(
-      filter(p => !!p),
-      switchMap(polygon => this.asfApiService.validate(polygon)),
-      map(resp => {
-        if (resp.error) {
-          const { report, type } = resp.error;
-
-          this.mapService.setDrawStyle(models.DrawPolygonStyle.INVALID);
-          this.snackBar.open(
-            report, 'INVALID POLYGON',
-            { duration: 4000, }
-          );
-
-          return;
-        } else {
-          this.mapService.setDrawStyle(models.DrawPolygonStyle.VALID);
-
-          const repairs = resp.repairs
-            .filter(repair =>
-              repair.type !== models.PolygonRepairTypes.ROUND
-            );
-
-          if (repairs.length === 0) {
-            return resp.wkt;
-          }
-
-          const { report, type }  = resp.repairs.pop();
-
-          this.snackBar.open(
-            report, type,
-            { duration: 4000, }
-          );
-
-          const features = this.wktService.wktToFeature(
-            resp.wkt,
-            this.mapService.epsg()
-          );
-
-          this.mapService.setDrawFeature(features);
-        }
-      }),
-      catchError((val, source) => source)
-    ).subscribe(_ => _);
   }
 }
